Check function information on the website of Taiwan Legislative API

Check function information on the website of Taiwan Legislative API

get_infos(x)

Arguments

x

The parameter should be `get_parlquestions`, `get_legislators`, `get_executive_response`, `get_bills`, or `get_meetings`

See also

Examples

get_infos("get_parlquestions")
#> $page_info
#> # A tibble: 12 × 2
#>    `資料集名稱：質詢事項(本院委員質詢部分)` 資料集描述                          
#>    <chr>                                    <chr>                               
#>  1 資料目錄                                 "質詢類"                            
#>  2 資料集說明                               "提供議事日程本院委員之質詢事項資訊…
#>  3 資料提供者                               "立法院開放資料服務平台系統(資訊處)"
#>  4 欄位說明                                 "term：屆別,sessionPeriod：會期,ses…
#>  5 資料集預覽                               "HTML預覽<aclass=\"btn_yellow\"href…
#>  6 資料集使用                               "使用方式及列表"                    
#>  7 API說明                                  "API提供參數分別為屆別(term)、會期(…
#>  8 連結說明                                 "https://data.ly.gov.tw/odw/ID6Acti…
#>  9 更新頻率                                 "每日01時55分"                      
#> 10 資料筆數                                 "9509"                              
#> 11 建立時間                                 "2014-08-18MonAug1806:00:00CST2014" 
#> 12 更新時間                                 "2022-10-25TueOct2501:57:10CST2022" 
#> 
#> $reference_url
#> [1] "https://data.ly.gov.tw/getds.action?id=6"
#> 

get_infos("get_legislators")
#> $page_info
#> # A tibble: 12 × 2
#>    `資料集名稱：歷屆委員資料` 資料集描述                                        
#>    <chr>                      <chr>                                             
#>  1 資料目錄                   "委員類"                                          
#>  2 資料集說明                 "提供委員基本資料。(自第2屆起)"                   
#>  3 資料提供者                 "立法院全球資訊網(資訊處)"                        
#>  4 欄位說明                   "term：屆別,name：委員姓名,ename：委員英文姓名,se…
#>  5 資料集預覽                 "HTML預覽<aclass=\"btn_yellow\"href='/openDataset…
#>  6 資料集使用                 "使用方式及列表"                                  
#>  7 API說明                    "API提供參數分別為屆別(term)、委員姓名(name)、性… 
#>  8 連結說明                   "https://data.ly.gov.tw/odw/ID16Action.action?nam…
#>  9 更新頻率                   "每日03時05分"                                    
#> 10 資料筆數                   "1531"                                            
#> 11 建立時間                   "2014-12-18ThuDec1810:30:58CST2014"               
#> 12 更新時間                   "2023-01-21SatJan2103:07:12CST2023"               
#> 
#> $reference_url
#> [1] "https://data.ly.gov.tw/getds.action?id=16"
#> 

get_infos("get_executive_response")
#> $page_info
#> # A tibble: 12 × 2
#>    `資料集名稱：行政院答復` 資料集描述                                          
#>    <chr>                    <chr>                                               
#>  1 資料目錄                 "質詢類"                                            
#>  2 資料集說明               "提供公報質詢事項行政院答復資訊。(自第8屆第1會期起)"
#>  3 資料提供者               "立法院開放資料服務平台系統(資訊處)"                
#>  4 欄位說明                 "term：屆別,sessionPeriod：會期,sessionTimes：會次,…
#>  5 資料集預覽               "HTML預覽<aclass=\"btn_yellow\"href='/openDatasetJs…
#>  6 資料集使用               "使用方式及列表"                                    
#>  7 API說明                  "API提供參數分別為屆別(term)、會期(sessionPeriod)、…
#>  8 連結說明                 "https://data.ly.gov.tw/odw/ID2Action.action?term=1…
#>  9 更新頻率                 "每日02時00分"                                      
#> 10 資料筆數                 "252040"                                            
#> 11 建立時間                 "2014-08-21ThuAug2103:00:00CST2014"                 
#> 12 更新時間                 "2023-01-24TueJan2402:40:22CST2023"                 
#> 
#> $reference_url
#> [1] "https://data.ly.gov.tw/getds.action?id=2"
#> 

get_infos("get_bills")
#> $page_info
#> [1] "￭ 本項提供「法律提案」之API供民眾使用，可自行下載法律提案之相關資訊，欄位包含：提案日期、提案名稱、屆別、會期、會次、主提案及連署提案等。"
#> [2] "￭ API提供參數計有3種，分別為提案日期、主提案人、檔案類型，其中提案日期與檔案類型為必填參數，檔案類型可為CSV、JSON、XML。"                 
#> [3] "￭ 範例說明(建議Chrome瀏覽器)：https://www.ly.gov.tw/WebAPI/LegislativeBill.aspx?from=1050201&to=1050531&proposer=委員姓名&mode=xml"       
#> 
#> $reference_url
#> [1] "https://www.ly.gov.tw/Pages/List.aspx?nodeid=153"
#> 

get_infos("get_meetings")
#> $page_info
#> [1] "￭ 本項提供「委員發言」之API供民眾使用，可自行下載委員發言之相關資訊，欄位包含：會議日期、會議狀態、會議名稱、會議事由、主辦單位及委員發言名單等。"
#> [2] "￭ API提供參數計有3種，分別為會議日期、主辦單位、檔案類型，其中會議日期與檔案類型為必填參數，檔案類型可為CSV、JSON、XML。"                         
#> [3] "￭ 範例說明(建議Chrome瀏覽器：https://www.ly.gov.tw/WebAPI/LegislativeSpeech.aspx?from=1061002&to=1061006&meeting_unit=內政委員會&mode=XML"        
#> 
#> $reference_url
#> [1] "https://www.ly.gov.tw/Pages/List.aspx?nodeid=154"
#>