Check Session Periods in Each Year (Minguo Calendar) 檢查每年會期 (民國曆)

Examines session periods in Taiwan Minguo calendar (檢查每年會期民國曆)

Usage

review_session_info(term)

review_session_info(term)

Arguments

term

numeric The term number (1-11) of Legislative Yuan

Value

dataframe

A dataframe containing session information

Details

review_session_info produces a dataframe, displaying each session period in year formatted in Minguo (Taiwan) calendar.

Check Session Periods in Each Year (Minguo Calendar)

The review_session_info function produces a dataframe, displaying each session period in year formatted in Minguo calendar. This implementation uses system curl command with –insecure option to bypass SSL/TLS issues.

See also

Author

David Liao (davidycliao@gmail.com)

Examples

# Show the session information for the 7th Legislative Yuan term periods in ROC calendar year
review_session_info(7)
#> Downloading data from Legislative Yuan website...
#> Session information retrieved successfully.
#> # A tibble: 14 × 2
#>    屆期會期                會議起迄日         
#>    <chr>                   <chr>              
#>  1 第7屆第8會期第1次臨時會 101/01/19~101/01/20
#>  2 第7屆第7會期            100/02/22~100/06/14
#>  3 第7屆第5會期第2次臨時會 99/08/17~99/08/30  
#>  4 第7屆第5會期            99/02/23~99/06/08  
#>  5 第7屆第4會期            98/09/18~99/01/12  
#>  6 第7屆第3會期            98/02/20~98/06/16  
#>  7 第7屆第2會期            97/09/19~98/01/13  
#>  8 第7屆第8會期            100/09/16~100/12/14
#>  9 第7屆第6會期            99/09/24~100/01/12 
#> 10 第7屆第5會期第1次臨時會 99/07/08~99/07/14  
#> 11 第7屆第4會期第1次臨時會 99/01/18~99/01/19  
#> 12 第7屆第3會期第1次臨時會 98/08/25~98/08/27  
#> 13 第7屆第2會期第1次臨時會 98/01/15~98/01/16  
#> 14 第7屆第1會期            97/02/22~97/07/18  
if (FALSE) { # \dontrun{
# Show the session information for the 7th Legislative Yuan term periods in ROC calendar year
review_session_info(7)
} # }