Check for the Website Availability I

This function checks the availability of a specified website by trying to read the first line of the site's content.

Usage

website_availability(site = "https://data.ly.gov.tw/index.action")

Arguments

site

A website URL to check. Default is "https://data.ly.gov.tw/index.action".

See also