Retrieving the records of bill sponsor and co-sponsor via Taiwan Legislative Yuan API

Retrieving the records of bill sponsor and co-sponsor via Taiwan Legislative Yuan API

get_bills(start_date = NULL, end_date = NULL, proposer = NULL, verbose = TRUE)

Arguments

start_date

Requesting meeting records starting from the date. A double represents a date in ROC Taiwan format. If a double is used, it should specify as Taiwan calendar format, e.g. 1090110.

end_date

Requesting meeting records ending from the date. A double represents a date in ROC Taiwan format.If a double is used, it should specify as Taiwan calendar format, e.g. 1090110.

proposer

The default value is NULL, which means all bill records are included between the starting date and the ending date.

verbose

The default value is TRUE, displaying the description of data retrieved in number, url and computing time.

Value

A tibble contains date, term, name, sessionPeriod, sessionTimes, billName, billProposer, billCosignatory, billStatus, date_ad

See also

Examples

get_bills(1060120, 1060220)
#>  Retrieved URL: 
#>  https://www.ly.gov.tw/WebAPI/LegislativeBill.aspx?from=1060120&to=1060220&proposer=&mode=json 
#>  Retrieved Bill Sponsor(s):  
#>  Retrieved date between: 2017-01-20 and 2017-02-20 
#>  Retrieved Num: 81 
#> $title
#> [1] "the records of bill sponsor and co-sponsor"
#> 
#> $query_time
#> [1] "2023-01-19 03:00:07 UTC"
#> 
#> $retrieved_number
#> [1] 81
#> 
#> $proposer
#> NULL
#> 
#> $start_date_ad
#> [1] "2017-01-20"
#> 
#> $end_date_ad
#> [1] "2017-02-20"
#> 
#> $start_date
#> [1] 1060120
#> 
#> $end_date
#> [1] 1060220
#> 
#> $url
#> [1] "https://www.ly.gov.tw/WebAPI/LegislativeBill.aspx?from=1060120&to=1060220&proposer=&mode=json"
#> 
#> $data
#> # A tibble: 81 × 9
#>    date    term  sessionPer…¹ sessi…² billN…³ billP…⁴ billC…⁵ billS…⁶ date_ad   
#>    <chr>   <chr> <chr>        <chr>   <chr>   <chr>   <chr>   <chr>   <date>    
#>  1 1060217 09    03           01      公教人… 民進黨… "曾銘…  ""      2017-02-17
#>  2 1060217 09    03           01      公務人… 民進黨… "曾銘…  ""      2017-02-17
#>  3 1060217 09    03           01      公務人… 民進黨… "曾銘…  ""      2017-02-17
#>  4 1060217 09    03           01      政務人… 民進黨… "曾銘…  ""      2017-02-17
#>  5 1060217 09    03           01      學校教… 民進黨… "曾銘…  ""      2017-02-17
#>  6 1060217 09    03           01      學校教… 民進黨… "曾銘…  ""      2017-02-17
#>  7 1060217 09    03           01      陸海空… 民進黨… "曾銘…  ""      2017-02-17
#>  8 1060217 09    03           01      軍人撫… 民進黨… "曾銘…  ""      2017-02-17
#>  9 1060217 09    03           01      都市更… 江永昌… "莊瑞…  ""      2017-02-17
#> 10 1060217 09    03           01      公益勸… 劉櫂豪… "趙天…  ""      2017-02-17
#> # … with 71 more rows, and abbreviated variable names ¹​sessionPeriod,
#> #   ²​sessionTimes, ³​billName, ⁴​billProposer, ⁵​billCosignatory, ⁶​billStatus
#>