Retrieving the records of the bills 法律提案(API)

Retrieving the records of the bills 法律提案(API)

Usage

get_bills(start_date = NULL, end_date = NULL, proposer = NULL, verbose = TRUE)

Arguments

start_date

Requesting meeting records starting from the date. A double represents a date in ROC Taiwan format. If a double is used, it should specify as Taiwan calendar format, e.g. 1090110.

end_date

Requesting meeting records ending from the date. A double represents a date in ROC Taiwan format.If a double is used, it should specify as Taiwan calendar format, e.g. 1090110.

proposer

The default value is NULL, which means all bill records are included between the starting date and the ending date.

verbose

The default value is TRUE, displaying the description of data retrieved in number, url and computing time.

Value

A tibble contains date, term, name, sessionPeriod, sessionTimes, billName, billProposer, billCosignatory, billStatus, date_ad

See also

Examples

## query bill records by a period of the dates in Taiwan ROC calender format
## 輸入「中華民國民年」下載立法委員提案資料
get_bills(start_date = 1060120, end_date = 1070310)
#>  Retrieved URL: 
#>  https://www.ly.gov.tw/WebAPI/LegislativeBill.aspx?from=1060120&to=1070310&proposer=&mode=json 
#>  Retrieved Bill Sponsor(s):  
#>  Retrieved date between: 2017-01-20 and 2018-03-10 
#>  Retrieved Num: 1589 
#> $title
#> [1] "the records of bill sponsor and co-sponsor"
#> 
#> $query_time
#> [1] "2023-01-30 16:52:38 UTC"
#> 
#> $retrieved_number
#> [1] 1589
#> 
#> $proposer
#> NULL
#> 
#> $start_date_ad
#> [1] "2017-01-20"
#> 
#> $end_date_ad
#> [1] "2018-03-10"
#> 
#> $start_date
#> [1] 1060120
#> 
#> $end_date
#> [1] 1070310
#> 
#> $url
#> [1] "https://www.ly.gov.tw/WebAPI/LegislativeBill.aspx?from=1060120&to=1070310&proposer=&mode=json"
#> 
#> $variable_names
#> [1] "date"            "term"            "sessionPeriod"  
#> [4] "sessionTimes"    "billName"        "billProposer"   
#> [7] "billCosignatory" "billStatus"      "date_ad"        
#> 
#> $manual_info
#> [1] "https://www.ly.gov.tw/Pages/List.aspx?nodeid=153"
#> 
#> $data
#> # A tibble: 1,589 × 9
#>    date   term  sessi…¹ sessi…² billN…³ billP…⁴ billC…⁵ billS…⁶ date_ad   
#>    <chr>  <chr> <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <date>    
#>  1 10703… 09    05      03      公職人… 鄭寶清… "黃秀…  ""      2018-03-09
#>  2 10703… 09    05      03      電信管… 鄭寶清  "莊瑞…  ""      2018-03-09
#>  3 10703… 09    05      03      國民教… 親民黨… "莊瑞…  ""      2018-03-09
#>  4 10703… 09    05      03      中央行… 親民黨… "莊瑞…  ""      2018-03-09
#>  5 10703… 09    05      03      行政院… 親民黨… "莊瑞…  ""      2018-03-09
#>  6 10703… 09    05      03      行政院… 親民黨… "莊瑞…  ""      2018-03-09
#>  7 10703… 09    05      03      經濟部… 親民黨… "莊瑞…  ""      2018-03-09
#>  8 10703… 09    05      03      姓名條… 鄭天財  "徐榛…  ""      2018-03-09
#>  9 10703… 09    05      03      勞工保… 鄭天財  "曾銘…  ""      2018-03-09
#> 10 10703… 09    05      03      全民健… 鄭天財  "徐榛…  ""      2018-03-09
#> # … with 1,579 more rows, and abbreviated variable names ¹​sessionPeriod,
#> #   ²​sessionTimes, ³​billName, ⁴​billProposer, ⁵​billCosignatory,
#> #   ⁶​billStatus
#> 

## query bill records by a period of the dates in Taiwan ROC calender format
## and a specific legislator
## 輸入「中華民國民年」與「指定立法委員」下載立法委員提案資料
get_bills(start_date = 1060120, end_date = 1070310,  proposer = "孔文吉")
#>  Retrieved URL: 
#>  https://www.ly.gov.tw/WebAPI/LegislativeBill.aspx?from=1060120&to=1070310&proposer=孔文吉&mode=json 
#>  Retrieved Bill Sponsor(s):  孔文吉 
#>  Retrieved date between: 2017-01-20 and 2018-03-10 
#>  Retrieved Num: 9 
#> $title
#> [1] "the records of bill sponsor and co-sponsor"
#> 
#> $query_time
#> [1] "2023-01-30 16:52:38 UTC"
#> 
#> $retrieved_number
#> [1] 9
#> 
#> $proposer
#> [1] "孔文吉"
#> 
#> $start_date_ad
#> [1] "2017-01-20"
#> 
#> $end_date_ad
#> [1] "2018-03-10"
#> 
#> $start_date
#> [1] 1060120
#> 
#> $end_date
#> [1] 1070310
#> 
#> $url
#> [1] "https://www.ly.gov.tw/WebAPI/LegislativeBill.aspx?from=1060120&to=1070310&proposer=孔文吉&mode=json"
#> 
#> $variable_names
#> [1] "date"            "term"            "sessionPeriod"  
#> [4] "sessionTimes"    "billName"        "billProposer"   
#> [7] "billCosignatory" "billStatus"      "date_ad"        
#> 
#> $manual_info
#> [1] "https://www.ly.gov.tw/Pages/List.aspx?nodeid=153"
#> 
#> $data
#> # A tibble: 9 × 9
#>   date    term  sessi…¹ sessi…² billN…³ billP…⁴ billC…⁵ billS…⁶ date_ad   
#>   <chr>   <chr> <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <date>    
#> 1 1061222 09    04      14      山坡地… 徐榛蔚… "許毓…  ""      2017-12-22
#> 2 1061208 09    04      12      客家基… 吳志揚… "顏寬…  ""      2017-12-08
#> 3 1061006 09    04      03      勞動基… 鄭天財… "陳宜…  ""      2017-10-06
#> 4 1060922 09    04      01      野生動… 高潞．… "洪慈…  ""      2017-09-22
#> 5 1060428 09    03      11      前瞻基… 孔文吉  "曾銘…  ""      2017-04-28
#> 6 1060421 09    03      10      轉型正… 國民黨… "林俊…  ""      2017-04-21
#> 7 1060317 09    03      05      就業服… 孔文吉… "曾銘…  ""      2017-03-17
#> 8 1060317 09    03      05      原住民… 孔文吉  "許毓…  ""      2017-03-17
#> 9 1060317 09    03      05      公益彩… 孔文吉… "簡東…  ""      2017-03-17
#> # … with abbreviated variable names ¹​sessionPeriod, ²​sessionTimes,
#> #   ³​billName, ⁴​billProposer, ⁵​billCosignatory, ⁶​billStatus
#> 

## query bill records by a period of the dates in Taiwan ROC calender format
## and multiple legislators
## 輸入「中華民國民年」與「指定多個立法委員」下載立法委員提案資料
get_bills(start_date = 1060120, end_date = 1060510,  proposer = "孔文吉&鄭天財")
#>  Retrieved URL: 
#>  https://www.ly.gov.tw/WebAPI/LegislativeBill.aspx?from=1060120&to=1060510&proposer=孔文吉&鄭天財&mode=json 
#>  Retrieved Bill Sponsor(s):  孔文吉&鄭天財 
#>  Retrieved date between: 2017-01-20 and 2017-05-10 
#>  Retrieved Num: 5 
#> $title
#> [1] "the records of bill sponsor and co-sponsor"
#> 
#> $query_time
#> [1] "2023-01-30 16:52:39 UTC"
#> 
#> $retrieved_number
#> [1] 5
#> 
#> $proposer
#> [1] "孔文吉&鄭天財"
#> 
#> $start_date_ad
#> [1] "2017-01-20"
#> 
#> $end_date_ad
#> [1] "2017-05-10"
#> 
#> $start_date
#> [1] 1060120
#> 
#> $end_date
#> [1] 1060510
#> 
#> $url
#> [1] "https://www.ly.gov.tw/WebAPI/LegislativeBill.aspx?from=1060120&to=1060510&proposer=孔文吉&鄭天財&mode=json"
#> 
#> $variable_names
#> [1] "date"            "term"            "sessionPeriod"  
#> [4] "sessionTimes"    "billName"        "billProposer"   
#> [7] "billCosignatory" "billStatus"      "date_ad"        
#> 
#> $manual_info
#> [1] "https://www.ly.gov.tw/Pages/List.aspx?nodeid=153"
#> 
#> $data
#> # A tibble: 5 × 9
#>   date    term  sessi…¹ sessi…² billN…³ billP…⁴ billC…⁵ billS…⁶ date_ad   
#>   <chr>   <chr> <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <date>    
#> 1 1060428 09    03      11      前瞻基… 孔文吉  "曾銘…  ""      2017-04-28
#> 2 1060421 09    03      10      轉型正… 國民黨… "林俊…  ""      2017-04-21
#> 3 1060317 09    03      05      就業服… 孔文吉… "曾銘…  ""      2017-03-17
#> 4 1060317 09    03      05      原住民… 孔文吉  "許毓…  ""      2017-03-17
#> 5 1060317 09    03      05      公益彩… 孔文吉… "簡東…  ""      2017-03-17
#> # … with abbreviated variable names ¹​sessionPeriod, ²​sessionTimes,
#> #   ³​billName, ⁴​billProposer, ⁵​billCosignatory, ⁶​billStatus
#>