The Records of Legislation and the Executives Proposals 委員及政府議案提案資訊

The Records of Legislation and the Executives Proposals 委員及政府議案提案資訊

Usage

get_bills_2(term = 8, session_period = NULL, verbose = TRUE)

Arguments

term

numeric or null. The data is only available from 8th term. The default value is 8. 參數必須為數值。資料從自第8屆起，預設值為8。

session_period

numeric or NULL. Available options for the session periods is: 1, 2, 3, 4, 5, 6, 7, and 8. The default is NULL. 參數必須為數值。

verbose

The default value is TRUE, displaying the description of data retrieved in number, url and computing time.

Value

list list, which contains:

title

the meeting records of cross-caucus session

query_time

the query time

retrieved_number

the number of observation

meeting_unit

the meeting unit

start_date_ad

the start date in POSIXct

end_date_ad

the end date in POSIXct

start_date

the start date in ROC Taiwan calendar

url

the retrieved json url

variable_names

the variables of the tibble dataframe

manual_info

the official manual, https://data.ly.gov.tw/getds.action?id=20; or use legisTaiwan::get_variable_info("get_bills_2")

data

a tibble dataframe, whose variables include:

sessionPeriod

會期

sessionTimes

會次

meetingTimes

臨時會會次

billNo

議案編號

billName

提案名稱

billOrg

提案單位/委員

billProposer

主提案人

billCosignatory

連署提案

billStatus

議案狀態

pdfUrl

關係文書pdf檔案下載位置

docUrl

關係文書doc檔案下載位置

selectTerm

屆別期別篩選條件

Details

get_bills_2 produces a list, which contains query_time, retrieved_number, meeting_unit, start_date_ad, end_date_ad, start_date, end_date, url, variable_names, manual_info and data. To retrieve the user manual and more information about the data frame, please use legisTaiwan::get_variable_info("get_bills_2").

Note

議事類: 提供委員及政府之議案提案資訊 (自第8屆第1會期起)。

See also

Examples

## query the Executives' answered response by term and the session period.
## 輸入「立委屆期」與「會期」下載「質詢事項 (行政院答復部分)」
get_bills_2(term = 8, session_period = 1)
#>  Retrieved URL: 
#>  https://data.ly.gov.tw/odw/ID20Action.action?term=08&sessionPeriod=01&sessionTimes=&meetingTimes=&billName=&billOrg=&billProposer=&billCosignatory=&fileType=json 
#>  Retrieved Term:  08 
#>  Retrieved Num:  1155 
#> $title
#> [1] "the records of the questions answered by the executives"
#> 
#> $query_time
#> [1] "2023-02-08 02:10:24 CET"
#> 
#> $retrieved_number
#> [1] 1155
#> 
#> $retrieved_term
#> [1] "08"
#> 
#> $url
#> [1] "https://data.ly.gov.tw/odw/ID20Action.action?term=08&sessionPeriod=01&sessionTimes=&meetingTimes=&billName=&billOrg=&billProposer=&billCosignatory=&fileType=json"
#> 
#> $variable_names
#>  [1] "term"            "sessionPeriod"   "sessionTimes"   
#>  [4] "meetingTimes"    "billNo"          "billName"       
#>  [7] "billOrg"         "billProposer"    "billCosignatory"
#> [10] "billStatus"      "pdfUrl"          "docUrl"         
#> [13] "selectTerm"     
#> 
#> $manual_info
#> [1] "https://data.ly.gov.tw/getds.action?id=2"
#> 
#> $data
#> # A tibble: 1,155 × 13
#>    term  sessionP…¹ sessi…² meeti…³ billNo billN…⁴ billOrg billP…⁵ billC…⁶
#>    <chr> <chr>      <chr>   <chr>   <chr>  <chr>   <chr>   <chr>   <chr>  
#>  1 08    01         02      ""      10102… 「公民… 本院委… "葉宜…  "潘孟… 
#>  2 08    01         02      ""      10102… 「兒童… 本院委… "葉宜…  "蔡其… 
#>  3 08    01         02      ""      10102… 「核子… 本院委… ""      ""     
#>  4 08    01         02      ""      10102… 「核子… 本院委… ""      ""     
#>  5 08    01         02      ""      10102… 「國家… 本院委… "潘孟…  "葉宜… 
#>  6 08    01         02      ""      10102… 「有線… 本院委… "潘孟…  "陳亭… 
#>  7 08    01         02      ""      10102… 「自來… 本院委… "李應…  "蔡煌… 
#>  8 08    01         02      ""      10102… 「就業… 本院委… "李應…  "蔡煌… 
#>  9 08    01         02      ""      10102… 「行政… 本院委… "李應…  "蔡煌… 
#> 10 08    01         02      ""      10102… 「立法… 本院委… ""      ""     
#> # … with 1,145 more rows, 4 more variables: billStatus <chr>,
#> #   pdfUrl <chr>, docUrl <chr>, selectTerm <chr>, and abbreviated
#> #   variable names ¹​sessionPeriod, ²​sessionTimes, ³​meetingTimes,
#> #   ⁴​billName, ⁵​billProposer, ⁶​billCosignatory
#>