Retrieving the records of legislators and the government proposals 議案提案

Retrieving the records of legislators and the government proposals 議案提案

Usage

get_bills_2(term = NULL, session_period = NULL, verbose = TRUE)

Arguments

term

Requesting answered questions by term. The parameter should be set in a numeric format. The default value is 8. The data is only available from 8th term 參數必須為數值，資料從立法院第8屆開始計算。

session_period

session in the term. The session is between 1 and 8. session_period 參數必須為數值。

verbose

The default value is TRUE, displaying the description of data retrieved in number, url and computing time.

Value

A list object contains a tibble carrying the variables of term, sessionPeriod, sessionTimes, meetingTimes, eyNumber, lyNumber, subject, content, docUrl selectTerm.

See also

Examples

## query the Executives' answered response by term and the session period.
## 輸入「立委屆期」與「會期」下載「質詢事項 (行政院答復部分)」
get_bills_2(term = 8, session_period = 1)
#>  Retrieved URL: 
#>  https://data.ly.gov.tw/odw/ID20Action.action?term=08&sessionPeriod=01&sessionTimes=&meetingTimes=&billName=&billOrg=&billProposer=&billCosignatory=&fileType=json 
#>  Retrieved Term:  8 
#>  Retrieved Num:  1155 
#> $title
#> [1] "the records of the questions answered by the executives"
#> 
#> $query_time
#> [1] "2023-01-30 11:31:18 UTC"
#> 
#> $retrieved_number
#> [1] 1155
#> 
#> $retrieved_term
#> [1] 8
#> 
#> $url
#> [1] "https://data.ly.gov.tw/odw/ID20Action.action?term=08&sessionPeriod=01&sessionTimes=&meetingTimes=&billName=&billOrg=&billProposer=&billCosignatory=&fileType=json"
#> 
#> $variable_names
#>  [1] "term"            "sessionPeriod"   "sessionTimes"   
#>  [4] "meetingTimes"    "billNo"          "billName"       
#>  [7] "billOrg"         "billProposer"    "billCosignatory"
#> [10] "billStatus"      "pdfUrl"          "docUrl"         
#> [13] "selectTerm"     
#> 
#> $manual_info
#> [1] "https://data.ly.gov.tw/getds.action?id=2"
#> 
#> $data
#> # A tibble: 1,155 × 13
#>    term  sessionP…¹ sessi…² meeti…³ billNo billN…⁴ billOrg billP…⁵ billC…⁶
#>    <chr> <chr>      <chr>   <chr>   <chr>  <chr>   <chr>   <chr>   <chr>  
#>  1 08    01         02      ""      10102… 「公民… 本院委… "葉宜…  "潘孟… 
#>  2 08    01         02      ""      10102… 「兒童… 本院委… "葉宜…  "蔡其… 
#>  3 08    01         02      ""      10102… 「核子… 本院委… ""      ""     
#>  4 08    01         02      ""      10102… 「核子… 本院委… ""      ""     
#>  5 08    01         02      ""      10102… 「國家… 本院委… "潘孟…  "葉宜… 
#>  6 08    01         02      ""      10102… 「有線… 本院委… "潘孟…  "陳亭… 
#>  7 08    01         02      ""      10102… 「自來… 本院委… "李應…  "蔡煌… 
#>  8 08    01         02      ""      10102… 「就業… 本院委… "李應…  "蔡煌… 
#>  9 08    01         02      ""      10102… 「行政… 本院委… "李應…  "蔡煌… 
#> 10 08    01         02      ""      10102… 「立法… 本院委… ""      ""     
#> # … with 1,145 more rows, 4 more variables: billStatus <chr>,
#> #   pdfUrl <chr>, docUrl <chr>, selectTerm <chr>, and abbreviated
#> #   variable names ¹​sessionPeriod, ²​sessionTimes, ³​meetingTimes,
#> #   ⁴​billName, ⁵​billProposer, ⁶​billCosignatory
#> 

get_bills_2(term = 8, session_period = 4)
#>  Retrieved URL: 
#>  https://data.ly.gov.tw/odw/ID20Action.action?term=08&sessionPeriod=04&sessionTimes=&meetingTimes=&billName=&billOrg=&billProposer=&billCosignatory=&fileType=json 
#>  Retrieved Term:  8 
#>  Retrieved Num:  1468 
#> $title
#> [1] "the records of the questions answered by the executives"
#> 
#> $query_time
#> [1] "2023-01-30 11:31:20 UTC"
#> 
#> $retrieved_number
#> [1] 1468
#> 
#> $retrieved_term
#> [1] 8
#> 
#> $url
#> [1] "https://data.ly.gov.tw/odw/ID20Action.action?term=08&sessionPeriod=04&sessionTimes=&meetingTimes=&billName=&billOrg=&billProposer=&billCosignatory=&fileType=json"
#> 
#> $variable_names
#>  [1] "term"            "sessionPeriod"   "sessionTimes"   
#>  [4] "meetingTimes"    "billNo"          "billName"       
#>  [7] "billOrg"         "billProposer"    "billCosignatory"
#> [10] "billStatus"      "pdfUrl"          "docUrl"         
#> [13] "selectTerm"     
#> 
#> $manual_info
#> [1] "https://data.ly.gov.tw/getds.action?id=2"
#> 
#> $data
#> # A tibble: 1,468 × 13
#>    term  sessionP…¹ sessi…² meeti…³ billNo billN…⁴ billOrg billP…⁵ billC…⁶
#>    <chr> <chr>      <chr>   <chr>   <chr>  <chr>   <chr>   <chr>   <chr>  
#>  1 08    04         02      ""      10205… 「加值… 本院委… 陳節如… 姚文智…
#>  2 08    04         02      ""      10205… 「發展… 本院委… 姚文智… 蔡其昌…
#>  3 08    04         02      ""      10205… 「行政… 本院委… 姚文智… 管碧玲…
#>  4 08    04         02      ""      10205… 「建築… 本院委… 姚文智… 李應元…
#>  5 08    04         02      ""      10205… 「都市… 本院委… 姚文智… 丁守中…
#>  6 08    04         02      ""      10205… 「都市… 本院委… 姚文智… 李應元…
#>  7 08    04         02      ""      10205… 「都市… 本院委… 林佳龍… 李應元…
#>  8 08    04         02      ""      10205… 「大學… 本院委… 林佳龍… 李應元…
#>  9 08    04         02      ""      10205… 「特殊… 本院委… 林佳龍… 林正二…
#> 10 08    04         02      ""      10205… 「銀行… 本院委… 林佳龍… 邱志偉…
#> # … with 1,458 more rows, 4 more variables: billStatus <chr>,
#> #   pdfUrl <chr>, docUrl <chr>, selectTerm <chr>, and abbreviated
#> #   variable names ¹​sessionPeriod, ²​sessionTimes, ³​meetingTimes,
#> #   ⁴​billName, ⁵​billProposer, ⁶​billCosignatory
#>