The Meeting Records of Cross-caucus Session 黨團協商資訊

The Meeting Records of Cross-caucus Session 黨團協商資訊

Usage

get_caucus_meetings(start_date = NULL, end_date = NULL, verbose = TRUE)

Arguments

start_date

character Must be formatted in Minguo (ROC) calendar with three forward slashes between year, month and day, e.g. "106/10/20".

end_date

character Must be formatted in Minguo (ROC) calendar with three forward slashes between year, month and day, e.g. "109/01/10".

verbose

logical, indicates whether get_caucus_meetings should print out detailed output when retrieving the data.

Value

list, which contains:

title

the meeting records of cross-caucus session

query_time

the query time

retrieved_number

the number of observation

meeting_unit

the meeting unit

start_date_ad

the start date in POSIXct

end_date_ad

the end date in POSIXct

start_date

the start date in ROC Taiwan calendar

url

the retrieved json url

variable_names

the variables of the tibble dataframe

manual_info

the official manual, https://data.ly.gov.tw/getds.action?id=8; or use legisTaiwan::get_variable_info("get_caucus_meetings")

data

a tibble dataframe, whose variables include:

comVolume

期

comBookId

冊別

term

屆別

sessionPeriod

會期

meetingTimes

臨時會會次

meetingDate

會議日期(民國年)

meetingName

會議名稱

subject

案由

pageStart

起始頁

pageEnd

結束頁

docUrl

檔案下載位置

htmlUrl

html網址

selectTerm

屆別期別篩選條件

Details

get_caucus_meetings produces a list, which contains title, query_time, retrieved_number, meeting_unit, start_date_ad, end_date_ad, start_date, end_date, url, variable_names, manual_info and data. \ifelsehtml\hrefhttps://lifecycle.r-lib.org/articles/stages.html#experimental\figurelifecycle-experimental.svgoptions: alt='[Experimental]'\strong[Experimental]

Note

To retrieve the user manual and more information about variable of the data frame, please use legisTaiwan::get_variable_info("get_caucus_meetings") or visit the API manual at https://data.ly.gov.tw/getds.action?id=8. 議事類:提供公報之黨團協商資訊 (自第8屆第1會期起)

See also

Author

David Liao (davidycliao@gmail.com)

Examples

## query the meeting records of cross-caucus session using a period of
## the dates in Taiwan ROC calender format with forward slash (/).
## 輸入「中華民國民年」下載「黨團協商」，輸入時間請依照該格式 "106/10/20"，
## 需有「正斜線」做隔開。
get_caucus_meetings(start_date = "106/10/20", end_date = "107/03/10")
#>  Retrieved URL: 
#>  https://data.ly.gov.tw/odw/ID8Action.action?comYear=&comVolume=&comBookId=&term=&sessionPeriod=&sessionTimes=&meetingTimes=&meetingDateS=106/10/20&meetingDateE=107/03/10&fileType=json 
#>  Retrieved date between: 2017-10-20 and 2018-03-10 
#>  Retrieved number: 30 
#> $title
#> [1] "the meeting records of cross-caucus session"
#> 
#> $query_time
#> [1] "2023-02-12 15:19:02 CET"
#> 
#> $retrieved_number
#> [1] 30
#> 
#> $start_date_ad
#> [1] "2017-10-20"
#> 
#> $end_date_ad
#> [1] "2018-03-10"
#> 
#> $start_date
#> [1] "106/10/20"
#> 
#> $end_date
#> [1] "107/03/10"
#> 
#> $url
#> [1] "https://data.ly.gov.tw/odw/ID8Action.action?comYear=&comVolume=&comBookId=&term=&sessionPeriod=&sessionTimes=&meetingTimes=&meetingDateS=106/10/20&meetingDateE=107/03/10&fileType=json"
#> 
#> $variable_names
#>  [1] "comYear"       "comVolume"     "comBookId"     "term"         
#>  [5] "sessionPeriod" "sessionTimes"  "meetingTimes"  "meetingDate"  
#>  [9] "meetingName"   "subject"       "pageStart"     "pageEnd"      
#> [13] "docUrl"        "htmlUrl"       "selectTerm"   
#> 
#> $manual_info
#> [1] "https://data.ly.gov.tw/getds.action?id=8"
#> 
#> $data
#> # A tibble: 30 × 15
#>    comYear comVolume comBo…¹ term  sessi…² sessi…³ meeti…⁴ meeti…⁵ meeti…⁶
#>    <chr>   <chr>     <chr>   <chr> <chr>   <chr>   <chr>   <chr>   <chr>  
#>  1 103     43        "四"    08    05      12      null    101053… 黨團協…
#>  2 103     43        "四"    08    05      12      null    102010… 黨團協…
#>  3 106     103       "下"    null  null    null    null    1061120 黨團協…
#>  4 106     108       ""      09    04      10      null    1061122 黨團協…
#>  5 106     108       ""      09    04      10      null    1061129 黨團協…
#>  6 106     113       "下"    null  null    null    null    1061130 黨團協…
#>  7 106     113       "下"    null  null    null    null    1061201 黨團協…
#>  8 106     116       "下"    09    04      12      null    1061208 黨團協…
#>  9 106     87        "下"    09    04      06      null    1061030 黨團協…
#> 10 106     87        "下"    09    04      null    null    1061031 立法院…
#> # … with 20 more rows, 6 more variables: subject <chr>, pageStart <chr>,
#> #   pageEnd <chr>, docUrl <chr>, htmlUrl <chr>, selectTerm <chr>, and
#> #   abbreviated variable names ¹​comBookId, ²​sessionPeriod, ³​sessionTimes,
#> #   ⁴​meetingTimes, ⁵​meetingDate, ⁶​meetingName
#>