Retrieving the meeting records of cross-caucus session 下載「黨團協商」資料

Retrieving the meeting records of cross-caucus session 下載「黨團協商」資料

Usage

get_caucus_meetings(start_date = NULL, end_date = NULL, verbose = TRUE)

Arguments

start_date

Requesting meeting records starting from the date. A double represents a date in ROC Taiwan format. If a double is used, it should specify as Taiwan calendar format, e.g. 109/01/10.

end_date

Requesting meeting records ending from the date. A double represents a date in ROC Taiwan format. If a double is used, it should specify as Taiwan calendar format, e.g. 109/01/20.

verbose

The default value is TRUE, displaying the description of data retrieved in number, url and computing time.

Value

A list carries a main tibble dataframe that contains comYear, comBookId , sessionPeriod, sessionTimes, htmlUrl, etc.

See also

Examples

## query the meeting records of cross-caucus session using a period of the dates
## in Taiwan ROC calender format with forward slash (/).
## 輸入「中華民國民年」下載「黨團協商」，輸入時間請依照該格式 "106/10/20"，需有「正斜線」做隔開。
get_caucus_meetings(start_date = "106/10/20", end_date = "107/03/10")
#>  Retrieved URL: 
#>  https://data.ly.gov.tw/odw/ID8Action.action?comYear=&comVolume=&comBookId=&term=&sessionPeriod=&sessionTimes=&meetingTimes=&meetingDateS=106/10/20&meetingDateE=107/03/10&fileType=json 
#>  Retrieved date between: 2017-10-20 and 2018-03-10 
#>  Retrieved number: 30 
#> $title
#> [1] "the meeting records of cross-caucus session"
#> 
#> $query_time
#> [1] "2023-01-27 20:49:26 GMT"
#> 
#> $retrieved_number
#> [1] 30
#> 
#> $start_date_ad
#> [1] "2017-10-20"
#> 
#> $end_date_ad
#> [1] "2018-03-10"
#> 
#> $start_date
#> [1] "106/10/20"
#> 
#> $end_date
#> [1] "107/03/10"
#> 
#> $url
#> [1] "https://data.ly.gov.tw/odw/ID8Action.action?comYear=&comVolume=&comBookId=&term=&sessionPeriod=&sessionTimes=&meetingTimes=&meetingDateS=106/10/20&meetingDateE=107/03/10&fileType=json"
#> 
#> $variable_names
#>  [1] "comYear"       "comVolume"     "comBookId"     "term"         
#>  [5] "sessionPeriod" "sessionTimes"  "meetingTimes"  "meetingDate"  
#>  [9] "meetingName"   "subject"       "pageStart"     "pageEnd"      
#> [13] "docUrl"        "htmlUrl"       "selectTerm"   
#> 
#> $manual_info
#> [1] "https://data.ly.gov.tw/getds.action?id=8"
#> 
#> $data
#> # A tibble: 30 × 15
#>    comYear comVolume comBo…¹ term  sessi…² sessi…³ meeti…⁴ meeti…⁵ meeti…⁶
#>    <chr>   <chr>     <chr>   <chr> <chr>   <chr>   <chr>   <chr>   <chr>  
#>  1 103     43        "四"    08    05      12      null    101053… 黨團協…
#>  2 103     43        "四"    08    05      12      null    102010… 黨團協…
#>  3 106     103       "下"    null  null    null    null    1061120 黨團協…
#>  4 106     108       ""      09    04      10      null    1061122 黨團協…
#>  5 106     108       ""      09    04      10      null    1061129 黨團協…
#>  6 106     113       "下"    null  null    null    null    1061130 黨團協…
#>  7 106     113       "下"    null  null    null    null    1061201 黨團協…
#>  8 106     116       "下"    09    04      12      null    1061208 黨團協…
#>  9 106     87        "下"    09    04      06      null    1061030 黨團協…
#> 10 106     87        "下"    09    04      null    null    1061031 立法院…
#> # … with 20 more rows, 6 more variables: subject <chr>, pageStart <chr>,
#> #   pageEnd <chr>, docUrl <chr>, htmlUrl <chr>, selectTerm <chr>, and
#> #   abbreviated variable names ¹​comBookId, ²​sessionPeriod, ³​sessionTimes,
#> #   ⁴​meetingTimes, ⁵​meetingDate, ⁶​meetingName
#>