The Records of Reviewed Items in the Committees 委員會會議審查之議案項目

The Records of Reviewed Items in the Committees 委員會會議審查之議案項目

Usage

get_committee_record(term = 8, session_period = NULL, verbose = TRUE)

Arguments

term

numeric or null. The data is only available from 8th term. The default value is 8. 參數必須為數值。資料從自第8屆起，預設值為8。

session_period

integer, numeric or NULL. Available options for the session is: 1, 2, 3, 4, 5, 6, 7, and 8. The default is NULL. 參數必須為數值。

verbose

logical, indicates whether get_executive_response should print out detailed output when retrieving the data. The default is TRUE

Value

list contains:

title

the records of the questions answered by the executives

query_time

the queried time

retrieved_number

the total number of observations

retrieved_term

the queried term

url

the retrieved json url

variable_names

the variables of the tibble dataframe

manual_info

the official manual, https://data.ly.gov.tw/getds.action?id=46; or use legisTaiwan::get_variable_info("get_committee_record")

data

a tibble dataframe, whose variables include:

sessionPeriod

會期

meetingNo

會議編號

billNo

議案編號

selectTerm

屆別期別篩選條件

Details

get_committee_record produces a list, which contains title, query_time, retrieved_number, retrieved_term, url, variable_names, manual_info and data. To retrieve the user manual and more information about the data frame, please use legisTaiwan::get_variable_info("get_committee_record").

Note

議事類: 提供委員會會議審查之議案項目 (自第8屆第1會期起)。

See also

Examples

## query the committee record by term and the session period.
## 輸入「立委屆期」與「會期」下載「委員會審議之議案」
get_committee_record(term = 8, session_period = 1)
#>  Retrieved URL: 
#>  https://data.ly.gov.tw/odw/ID46Action.action?term=08&sessionPeriod=01&sessionTimes=01&meetingTimes=&fileType=json 
#>  Retrieved Term:  08 
#>  Retrieved Num:  613 
#> $title
#> [1] "the records of reviewed items in the committees"
#> 
#> $query_time
#> [1] "2023-02-07 22:42:51 CET"
#> 
#> $retrieved_number
#> [1] 613
#> 
#> $retrieved_term
#> [1] "08"
#> 
#> $url
#> [1] "https://data.ly.gov.tw/odw/ID46Action.action?term=08&sessionPeriod=01&sessionTimes=01&meetingTimes=&fileType=json"
#> 
#> $variable_names
#> [1] "term"          "sessionPeriod" "meetingNo"     "billNo"       
#> [5] "billName"      "docUrl"        "selectTerm"   
#> 
#> $manual_info
#> [1] "https://data.ly.gov.tw/getds.action?id=46"
#> 
#> $data
#> # A tibble: 613 × 7
#>    term  sessionPeriod meetingNo billNo           billName  docUrl selec…¹
#>    <chr> <chr>         <chr>     <chr>            <chr>     <chr>  <chr>  
#>  1 08    01            201203098 0990901071000100 國家通訊… https… all    
#>  2 08    01            201203098 0990903071100100 交通部函… https… all    
#>  3 08    01            201203098 1000525070100100 交通部函… https… all    
#>  4 08    01            201203098 1000602070100400 國家通訊… https… all    
#>  5 08    01            201203098 1000902070100400 國家通訊… https… all    
#>  6 08    01            201203098 1000907070100500 交通部函… https… all    
#>  7 08    01            201203098 1010224070202300 本院委員… https… all    
#>  8 08    01            201203099 1010224070200800 本院委員… https… all    
#>  9 08    01            201203099 1010224070200900 本院親民… https… all    
#> 10 08    01            201203099 1010224070202100 本院委員… https… all    
#> # … with 603 more rows, and abbreviated variable name ¹​selectTerm
#>