Retrieving the records of the questions answered by the executives 提供公報質詢事項行政院答復資訊

Retrieving the records of the questions answered by the executives 提供公報質詢事項行政院答復資訊

Usage

get_executive_response(term = NULL, session_period = NULL, verbose = TRUE)

Arguments

term

integer, numeric or null. The default is 8. The data is only available from 8th term. 參數必須為數值，資料從自第8屆第1會期起。

session_period

integer, numeric or NULL. Available options for the session is: 1, 2, 3, 4, 5, 6, 7, and 8. The default is NULL. 參數必須為數值。

verbose

logical, indicates whether get_executive_response should print out detailed output when retrieving the data. The default is TRUE

Value

list contains:

title

the records of the questions answered by the executives

query_time

the queried time

retrieved_number

the total number of observations

retrieved_term

the queried term

url

the retrieved json url

variable_names

the variables of the tibble dataframe

manual_info

the offical manual

data

a tibble dataframe , whose variables include: term, sessionPeriod, sessionTimes, meetingTimes, eyNumber, lyNumber, subject, content, docUrl, and selectTerm

Details

get_executive_response produces a list, which contains title, query_time, retrieved_number, retrieved_term, url, variable_names, manual_info and data.

See also

Examples

## query the Executives' answered response by term and the session period.
## 輸入「立委屆期」與「會期」下載「行政院答復」
get_executive_response(term = 8, session_period = 1)
#>  Retrieved URL: 
#>  https://data.ly.gov.tw/odw/ID2Action.action?term=08&sessionPeriod=01&sessionTimes=&item=&fileType=json 
#>  Retrieved Term:  8 
#>  Retrieved Num:  1065 
#> $title
#> [1] "the records of the questions answered by the executives"
#> 
#> $query_time
#> [1] "2023-02-01 22:03:08 GMT"
#> 
#> $retrieved_number
#> [1] 1065
#> 
#> $retrieved_term
#> [1] 8
#> 
#> $url
#> [1] "https://data.ly.gov.tw/odw/ID2Action.action?term=08&sessionPeriod=01&sessionTimes=&item=&fileType=json"
#> 
#> $variable_names
#>  [1] "term"          "sessionPeriod" "sessionTimes"  "meetingTimes" 
#>  [5] "eyNumber"      "lyNumber"      "subject"       "content"      
#>  [9] "docUrl"        "selectTerm"   
#> 
#> $manual_info
#> [1] "https://data.ly.gov.tw/getds.action?id=2"
#> 
#> $data
#> # A tibble: 1,065 × 10
#>    term  sessionP…¹ sessi…² meeti…³ eyNum…⁴ lyNum…⁵ subject content docUrl
#>    <chr> <chr>      <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr> 
#>  1 08    01         01      null    （行政… （立法… （一八… 盧委員… http:…
#>  2 08    01         01      null    （行政… （立法… （一八… 陳委員… http:…
#>  3 08    01         01      null    （行政… （立法… （一八… 王委員… http:…
#>  4 08    01         01      null    （行政… （立法… （一八… 有關王… http:…
#>  5 08    01         01      null    （行政… （立法… （一八… 陳委員… http:…
#>  6 08    01         01      null    （行政… （立法… （一八… 陳委員… http:…
#>  7 08    01         01      null    （行政… （立法… （一八… 蔡委員… http:…
#>  8 08    01         01      null    （行政… （立法… （一八… 一、有… http:…
#>  9 08    01         01      null    （行政… （立法… （一八… 蔡委員… http:…
#> 10 08    01         01      null    （行政… （立法… （一九… 蔡委員… http:…
#> # … with 1,055 more rows, 1 more variable: selectTerm <chr>, and
#> #   abbreviated variable names ¹​sessionPeriod, ²​sessionTimes,
#> #   ³​meetingTimes, ⁴​eyNumber, ⁵​lyNumber
#>