Retrieving legislator' demographic information and background 提供委員基本資料

Retrieving legislator' demographic information and background 提供委員基本資料

Usage

get_legislators(term = NULL, verbose = TRUE)

Arguments

term

numeric or null. The data is available from the 2nd term. 自第2屆起

verbose

logical, indicates whether get_meetings should print out detailed output when retrieving the data. The default is TRUE.

Value

list contains:

query_time

the queried time

queried_term

the queried term

url

the retrieved json url

variable_names

the variables of the tibble dataframe

manual_info

the offical manual

data

a tibble dataframe , whose variables include: term, name, ename, sex, party, partyGroup, areaName, committee, onboardDate, degree, picUrl, leaveFlag, leaveDate and leaveReason

Details

get_legislators produces a list, which contains query_time, queried_term, url, variable_names, manual_info and data.

See also

Examples

## To retrieve legislator' demographic from 8th
## 輸入「屆次」下載當屆立委資料
get_legislators(term = 8)
#>  Retrieved URL: 
#>  https://data.ly.gov.tw/odw/ID16Action.action?name=&sex=&party=&partyGroup=&areaName=&term=08=&fileType=json 
#>  Retrieved term:  8 
#> $title
#> [1] "the legislator's demographic information"
#> 
#> $query_time
#> [1] "2023-02-03 09:33:01 UTC"
#> 
#> $queried_term
#> [1] "8"
#> 
#> $url
#> [1] "https://data.ly.gov.tw/odw/ID16Action.action?name=&sex=&party=&partyGroup=&areaName=&term=08=&fileType=json"
#> 
#> $variable_names
#>  [1] "term"        "name"        "ename"       "sex"         "party"      
#>  [6] "partyGroup"  "areaName"    "committee"   "onboardDate" "degree"     
#> [11] "experience"  "picUrl"      "leaveFlag"   "leaveDate"   "leaveReason"
#> 
#> $manual_info
#> [1] "https://data.ly.gov.tw/getds.action?id=16"
#> 
#> $data
#> # A tibble: 124 × 15
#>    term  name   ename   sex   party party…¹ areaN…² commi…³ onboa…⁴ degree
#>    <chr> <chr>  <chr>   <chr> <chr> <chr>   <chr>   <chr>   <chr>   <chr> 
#>  1 08    丁守中 Ting S… 男    中國… 中國國… 台北市… 第8屆…  2012/0… 美國… 
#>  2 08    王金平 Wang J… 男    中國… 中國國… 全國不… 第8屆…  2012/0… 台灣… 
#>  3 08    王廷升 Wang T… 男    中國… 中國國… 花蓮縣… 第8屆…  2012/0… 美國… 
#>  4 08    王育敏 Wang Y… 女    中國… 中國國… 全國不… 第8屆…  2012/0… 東海… 
#>  5 08    王惠美 Wang H… 女    中國… 中國國… 彰化縣… 第8屆…  2012/0… 東海… 
#>  6 08    王進士 Wang C… 男    中國… 中國國… 屏東縣… 第8屆…  2012/0… 國立… 
#>  7 08    孔文吉 Kung W… 男    中國… 中國國… 山地原… 第8屆…  2012/0… 台灣… 
#>  8 08    尤美女 Yu Mei… 女    民主… 民主進… 全國不… 第8屆…  2012/0… 國立… 
#>  9 08    田秋堇 TienCh… 女    民主… 民主進… 全國不… 第8屆…  2012/0… 台灣… 
#> 10 08    江啟臣 Chiang… 男    中國… 中國國… 台中市… 第8屆…  2012/0… 美國… 
#> # … with 114 more rows, 5 more variables: experience <chr>, picUrl <chr>,
#> #   leaveFlag <chr>, leaveDate <chr>, leaveReason <chr>, and abbreviated
#> #   variable names ¹​partyGroup, ²​areaName, ³​committee, ⁴​onboardDate
#> 

## To retrieve legislator' demographic from all of terms
## 輸入「空白」下載所有立委資料
get_legislators()
#>  term is not defined...
#>  You are now requesting full data from the API. Please make sure your connectivity is stable until its completion.
#>  Retrieved URL: 
#>  https://data.ly.gov.tw/odw/ID16Action.action?name=&sex=&party=&partyGroup=&areaName=&term==&fileType=json 
#>  Retrieved term:  2 3 4 5 6 7 8 9 10 
#> $title
#> [1] "the legislator's demographic information"
#> 
#> $query_time
#> [1] "2023-02-03 09:33:03 UTC"
#> 
#> $queried_term
#> [1] "2 3 4 5 6 7 8 9 10"
#> 
#> $url
#> [1] "https://data.ly.gov.tw/odw/ID16Action.action?name=&sex=&party=&partyGroup=&areaName=&term==&fileType=json"
#> 
#> $variable_names
#>  [1] "term"        "name"        "ename"       "sex"         "party"      
#>  [6] "partyGroup"  "areaName"    "committee"   "onboardDate" "degree"     
#> [11] "experience"  "picUrl"      "leaveFlag"   "leaveDate"   "leaveReason"
#> 
#> $manual_info
#> [1] "https://data.ly.gov.tw/getds.action?id=16"
#> 
#> $data
#> # A tibble: 1,531 × 15
#>    term  name     ename sex   party party…¹ areaN…² commi…³ onboa…⁴ degree
#>    <chr> <chr>    <chr> <chr> <chr> <chr>   <chr>   <chr>   <chr>   <chr> 
#>  1 10    王定宇   Wang… 男    民主… 民主進… 臺南市… 第10屆… 2020/0… 國立… 
#>  2 10    孔文吉   Kung… 男    中國… 中國國… 山地原… 第10屆… 2020/0… 英國… 
#>  3 10    王美惠   Wang… 女    民主… 民主進… 嘉義市… 第10屆… 2020/0… 嘉義… 
#>  4 10    王婉諭   Wang… 女    時代… 時代力… 全國不… 第10屆… 2020/0… 苗栗… 
#>  5 10    伍麗華S… WuLi… 女    民主… 民主進… 山地原… 第10屆… 2020/0… 國立… 
#>  6 10    江永昌   Chia… 男    民主… 民主進… 新北市… 第10屆… 2020/0… 中和… 
#>  7 10    江啟臣   Chia… 男    中國… 中國國… 臺中市… 第10屆… 2020/0… 美國… 
#>  8 10    何志偉   Ho M… 男    民主… 民主進… 臺北市… 第10屆… 2020/0… 美國… 
#>  9 10    何欣純   Ho H… 女    民主… 民主進… 臺中市… 第10屆… 2020/0… 英國… 
#> 10 10    余天     YuTi… 男    民主… 民主進… 新北市… 第10屆… 2020/0… 新竹… 
#> # … with 1,521 more rows, 5 more variables: experience <chr>,
#> #   picUrl <chr>, leaveFlag <chr>, leaveDate <chr>, leaveReason <chr>,
#> #   and abbreviated variable names ¹​partyGroup, ²​areaName, ³​committee,
#> #   ⁴​onboardDate
#>