The Legislator' Demographic Information and Background 提供委員基本資料

The Legislator' Demographic Information and Background 提供委員基本資料

Usage

get_legislators(term = NULL, verbose = TRUE)

Arguments

term

numeric or NULL The data is available from the 2nd term.

verbose

logical, indicates whether get_meetings should print out detailed output when retrieving the data. The default is set to TRUE.

Value

list contains:

query_time

the queried time

queried_term

the queried term

url

the retrieved json url

variable_names

the variables of the tibble dataframe

manual_info

the official manual from https://data.ly.gov.tw/getds.action?id=16, or use legisTaiwan::get_variable_info("get_legislators")

data

a tibble dataframe, whose variables include:

name

委員姓名

ename

委員姓名

sex

性別

party

黨籍

partyGroup

黨團

committee

委員會

onboardDate

到職日(西元年)

degree

學歷

experience

經歷

picPath

照片位址

leaveFlag

離職日期(西元年)

leaveReason

離職原因

Details

get_legislators produces a list, which contains query_time, queried_term, url, variable_names, manual_info and data.

Note

To retrieve the user manual and more information about variable of the data frame, please use legisTaiwan::get_variable_info("get_legislators") or visit the API manual at https://data.ly.gov.tw/getds.action?id=16. 提供委員基本資料，最早資料可追溯至第2屆。

See also

Author

David Liao (davidycliao@gmail.com)

Examples

## query the Executives' answered response by term and the session period.
## 輸入「立委屆期」與「會期」下載「行政院答復」
get_executive_response(term = 8, session_period = 1)
#>  Retrieved URL: 
#>  https://data.ly.gov.tw/odw/ID2Action.action?term=08&sessionPeriod=01&sessionTimes=&item=&fileType=json 
#>  Retrieved Term:  08 
#>  Retrieved Num:  1065 
#> $title
#> [1] "the records of the questions answered by the executives"
#> 
#> $query_time
#> [1] "2023-02-12 15:07:41 CET"
#> 
#> $retrieved_number
#> [1] 1065
#> 
#> $retrieved_term
#> [1] "08"
#> 
#> $url
#> [1] "https://data.ly.gov.tw/odw/ID2Action.action?term=08&sessionPeriod=01&sessionTimes=&item=&fileType=json"
#> 
#> $variable_names
#>  [1] "term"          "sessionPeriod" "sessionTimes"  "meetingTimes" 
#>  [5] "eyNumber"      "lyNumber"      "subject"       "content"      
#>  [9] "docUrl"        "selectTerm"   
#> 
#> $manual_info
#> [1] "https://data.ly.gov.tw/getds.action?id=2"
#> 
#> $data
#> # A tibble: 1,065 × 10
#>    term  sessionP…¹ sessi…² meeti…³ eyNum…⁴ lyNum…⁵ subject content docUrl
#>    <chr> <chr>      <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr> 
#>  1 08    01         01      null    （行政… （立法… （一八… 盧委員… http:…
#>  2 08    01         01      null    （行政… （立法… （一八… 陳委員… http:…
#>  3 08    01         01      null    （行政… （立法… （一八… 王委員… http:…
#>  4 08    01         01      null    （行政… （立法… （一八… 有關王… http:…
#>  5 08    01         01      null    （行政… （立法… （一八… 陳委員… http:…
#>  6 08    01         01      null    （行政… （立法… （一八… 陳委員… http:…
#>  7 08    01         01      null    （行政… （立法… （一八… 蔡委員… http:…
#>  8 08    01         01      null    （行政… （立法… （一八… 一、有… http:…
#>  9 08    01         01      null    （行政… （立法… （一八… 蔡委員… http:…
#> 10 08    01         01      null    （行政… （立法… （一九… 蔡委員… http:…
#> # … with 1,055 more rows, 1 more variable: selectTerm <chr>, and
#> #   abbreviated variable names ¹​sessionPeriod, ²​sessionTimes,
#> #   ³​meetingTimes, ⁴​eyNumber, ⁵​lyNumber
#>