Retrieving legislator' demographics via Taiwan Legislative Yuan API 歷屆委員資料

Retrieving legislator' demographics via Taiwan Legislative Yuan API 歷屆委員資料

get_legislators(term = NULL, verbose = TRUE)

Arguments

term

Requesting answered questions from the term. The parameter should be set in a numeric format. The default value is 2. retrieved in number, url, and computing time.

verbose

The default is TRUE, which return the information of the return object.

Value

A list object contains a tibble carrying term, name, ename, sex, party, etc.

See also

Examples

## To retrieve legislator' demographic from 8th
## 輸入「屆次」下載當屆立委資料
get_legislators(term = 8)
#>  Retrieved URL: 
#>  https://data.ly.gov.tw/odw/ID16Action.action?name=&sex=&party=&partyGroup=&areaName=&term=08=&fileType=json 
#>  Retrieved term:  8 
#> $title
#> [1] "the legislator's demographic information"
#> 
#> $query_time
#> [1] "2023-01-23 18:06:56 UTC"
#> 
#> $queried_term
#> [1] "8"
#> 
#> $url
#> [1] "https://data.ly.gov.tw/odw/ID16Action.action?name=&sex=&party=&partyGroup=&areaName=&term=08=&fileType=json"
#> 
#> $variable_names
#>  [1] "term"        "name"        "ename"       "sex"         "party"      
#>  [6] "partyGroup"  "areaName"    "committee"   "onboardDate" "degree"     
#> [11] "experience"  "picUrl"      "leaveFlag"   "leaveDate"   "leaveReason"
#> 
#> $manual_info
#> [1] "https://data.ly.gov.tw/getds.action?id=16"
#> 
#> $data
#> # A tibble: 124 × 15
#>    term  name   ename sex   party party…¹ areaN…² commi…³ onboa…⁴ degree exper…⁵
#>    <chr> <chr>  <chr> <chr> <chr> <chr>   <chr>   <chr>   <chr>   <chr>  <chr>  
#>  1 08    丁守中 Ting… 男    中國… 中國國… 台北市… 第8屆…  2012/0… 美國…  中華民…
#>  2 08    王金平 Wang… 男    中國… 中國國… 全國不… 第8屆…  2012/0… 台灣…  高雄縣…
#>  3 08    王廷升 Wang… 男    中國… 中國國… 花蓮縣… 第8屆…  2012/0… 美國…  中華民…
#>  4 08    王育敏 Wang… 女    中國… 中國國… 全國不… 第8屆…  2012/0… 東海…  總統府…
#>  5 08    王惠美 Wang… 女    中國… 中國國… 彰化縣… 第8屆…  2012/0… 東海…  鹿鳴國…
#>  6 08    王進士 Wang… 男    中國… 中國國… 屏東縣… 第8屆…  2012/0… 國立…  第7屆… 
#>  7 08    孔文吉 Kung… 男    中國… 中國國… 山地原… 第8屆…  2012/0… 台灣…  行政院…
#>  8 08    尤美女 Yu M… 女    民主… 民主進… 全國不… 第8屆…  2012/0… 國立…  尤美女…
#>  9 08    田秋堇 Tien… 女    民主… 民主進… 全國不… 第8屆…  2012/0… 台灣…  台灣環…
#> 10 08    江啟臣 Chia… 男    中國… 中國國… 台中市… 第8屆…  2012/0… 美國…  行政院…
#> # … with 114 more rows, 4 more variables: picUrl <chr>, leaveFlag <chr>,
#> #   leaveDate <chr>, leaveReason <chr>, and abbreviated variable names
#> #   ¹​partyGroup, ²​areaName, ³​committee, ⁴​onboardDate, ⁵​experience
#> 

## To retrieve legislator' demographic from all of terms
## 輸入「空白」下載所有立委資料
get_legislators()
#>  Retrieved URL: 
#>  https://data.ly.gov.tw/odw/ID16Action.action?name=&sex=&party=&partyGroup=&areaName=&term==&fileType=json 
#>  Retrieved term:  2 3 4 5 6 7 8 9 10 
#> $title
#> [1] "the legislator's demographic information"
#> 
#> $query_time
#> [1] "2023-01-23 18:06:58 UTC"
#> 
#> $queried_term
#> [1] "2 3 4 5 6 7 8 9 10"
#> 
#> $url
#> [1] "https://data.ly.gov.tw/odw/ID16Action.action?name=&sex=&party=&partyGroup=&areaName=&term==&fileType=json"
#> 
#> $variable_names
#>  [1] "term"        "name"        "ename"       "sex"         "party"      
#>  [6] "partyGroup"  "areaName"    "committee"   "onboardDate" "degree"     
#> [11] "experience"  "picUrl"      "leaveFlag"   "leaveDate"   "leaveReason"
#> 
#> $manual_info
#> [1] "https://data.ly.gov.tw/getds.action?id=16"
#> 
#> $data
#> # A tibble: 1,531 × 15
#>    term  name   ename sex   party party…¹ areaN…² commi…³ onboa…⁴ degree exper…⁵
#>    <chr> <chr>  <chr> <chr> <chr> <chr>   <chr>   <chr>   <chr>   <chr>  <chr>  
#>  1 10    王定宇 Wang… 男    民主… 民主進… 臺南市… 第10屆… 2020/0… 國立…  2002年…
#>  2 10    孔文吉 Kung… 男    中國… 中國國… 山地原… 第10屆… 2020/0… 英國…  第六、…
#>  3 10    王美惠 Wang… 女    民主… 民主進… 嘉義市… 第10屆… 2020/0… 嘉義…  嘉義市…
#>  4 10    王婉諭 Wang… 女    時代… 時代力… 全國不… 第10屆… 2020/0… 苗栗…  新竹科…
#>  5 10    伍麗…  WuLi… 女    民主… 民主進… 山地原… 第10屆… 2020/0… 國立…  泰武國…
#>  6 10    江永昌 Chia… 男    民主… 民主進… 新北市… 第10屆… 2020/0… 中和…  公督盟…
#>  7 10    江啟臣 Chia… 男    中國… 中國國… 臺中市… 第10屆… 2020/0… 美國…  中華民…
#>  8 10    何志偉 Ho M… 男    民主… 民主進… 臺北市… 第10屆… 2020/0… 美國…  中華民…
#>  9 10    何欣純 Ho H… 女    民主… 民主進… 臺中市… 第10屆… 2020/0… 英國…  第九屆…
#> 10 10    余天   YuTi… 男    民主… 民主進… 新北市… 第10屆… 2020/0… 新竹…  現任第…
#> # … with 1,521 more rows, 4 more variables: picUrl <chr>, leaveFlag <chr>,
#> #   leaveDate <chr>, leaveReason <chr>, and abbreviated variable names
#> #   ¹​partyGroup, ²​areaName, ³​committee, ⁴​onboardDate, ⁵​experience
#>