Retrieving the spoken meeting records via Taiwan Legislative Yuan (The Legislature) API

Retrieving the spoken meeting records via Taiwan Legislative Yuan (The Legislature) API

get_meetings(
  start_date = NULL,
  end_date = NULL,
  meeting_unit = NULL,
  verbose = TRUE
)

Arguments

start_date

Requesting meeting records starting from the date. A double represents a date in ROC Taiwan format. If a double is used, it should specify as Taiwan calendar format, e.g. 1090110.

end_date

Requesting meeting records ending from the date. A double represents a date in ROC Taiwan format. If a double is used, it should specify as Taiwan calendar format, e.g. 1090110.

meeting_unit

The default is NULL, which include all meetings between the starting date and the ending date.

verbose

The default value is TRUE, displaying the description of data retrieved in number, url and computing time.

Value

A list carries a main tibble dataframe that contains the date, status, name, content and speakers.

See also

Examples

## query meeting records by a period of the dates in Taiwan ROC calender format
## 輸入「中華民國民年」下載會議審查資訊
get_meetings(1050120, 1050210)
#>  Retrieved URL: 
#>  https://www.ly.gov.tw/WebAPI/LegislativeSpeech.aspx?from=1050120&to=1050210&meeting_unit=&mode=json 
#>  Retrieved via : 
#>  Retrieved date between: 2016-01-20 and 2016-02-10 
#>  Retrieved number: 1 
#> $title
#> [1] "the spoken meeting records"
#> 
#> $query_time
#> [1] "2023-01-19 16:07:38 UTC"
#> 
#> $retrieved_number
#> [1] 1
#> 
#> $meeting_unit
#> NULL
#> 
#> $start_date_ad
#> [1] "2016-01-20"
#> 
#> $end_date_ad
#> [1] "2016-02-10"
#> 
#> $start_date
#> [1] 1050120
#> 
#> $end_date
#> [1] 1050210
#> 
#> $url
#> [1] "https://www.ly.gov.tw/WebAPI/LegislativeSpeech.aspx?from=1050120&to=1050210&meeting_unit=&mode=json"
#> 
#> $data
#> # A tibble: 1 × 7
#>   smeeting_date meeting_status meeting_name   meeti…¹ speec…² meeti…³ date_ad   
#>   <chr>         <chr>          <chr>          <chr>   <lgl>   <chr>   <date>    
#> 1 105/02/01     散會(15:55)    第9屆立法院預… 第9屆…  NA      院會    2016-02-01
#> # … with abbreviated variable names ¹​meeting_content, ²​speechers, ³​meeting_unit
#> 

## query meeting records by a period of the dates in Taiwan ROC calender format
## and a meeting
## 輸入「中華民國民年」與「審查會議或委員會名稱」下載會議審查資訊
get_meetings(start_date = 1060120, end_date = 1070310, meeting_unit = "內政委員會")
#>  Retrieved URL: 
#>  https://www.ly.gov.tw/WebAPI/LegislativeSpeech.aspx?from=1060120&to=1070310&meeting_unit=內政委員會&mode=json 
#>  Retrieved via : 內政委員會 
#>  Retrieved date between: 2017-01-20 and 2018-03-10 
#>  Retrieved number: 104 
#> $title
#> [1] "the spoken meeting records"
#> 
#> $query_time
#> [1] "2023-01-19 16:07:38 UTC"
#> 
#> $retrieved_number
#> [1] 104
#> 
#> $meeting_unit
#> [1] "內政委員會"
#> 
#> $start_date_ad
#> [1] "2017-01-20"
#> 
#> $end_date_ad
#> [1] "2018-03-10"
#> 
#> $start_date
#> [1] 1060120
#> 
#> $end_date
#> [1] 1070310
#> 
#> $url
#> [1] "https://www.ly.gov.tw/WebAPI/LegislativeSpeech.aspx?from=1060120&to=1070310&meeting_unit=內政委員會&mode=json"
#> 
#> $data
#> # A tibble: 104 × 7
#>    smeeting_date meeting_status meeting_name  meeti…¹ speec…² meeti…³ date_ad   
#>    <chr>         <chr>          <chr>         <chr>   <chr>   <chr>   <date>    
#>  1 107/03/08     散會(10:58)    立法院第9屆…  "選舉…   NA     內政委… 2018-03-08
#>  2 106/12/28     散會(17:05)    立法院第9屆…  "一、…   NA     內政委… 2017-12-28
#>  3 106/12/27     散會(17:36)    立法院第9屆…  "一、…   NA     內政委… 2017-12-27
#>  4 106/12/25     散會(12:50)    立法院第9屆…  "一、…  " 0001… 內政委… 2017-12-25
#>  5 106/12/25     散會(16:56)    立法院第9屆…  " "      NA     內政委… 2017-12-25
#>  6 106/12/22     會議結束       立法院朝野黨… "本院…   NA     朝野黨… 2017-12-22
#>  7 106/12/21     散會(16:46)    立法院第9屆…  "一、…  " 0001… 內政委… 2017-12-21
#>  8 106/12/20     休息           立法院第9屆…  "一、…  " 0001… 內政委… 2017-12-20
#>  9 106/12/18     散會(16:15)    立法院第9屆…  "一、…   NA     內政委… 2017-12-18
#> 10 106/12/14     散會(14:30)    立法院第9屆…  "一、…  " 0001… 內政委… 2017-12-14
#> # … with 94 more rows, and abbreviated variable names ¹​meeting_content,
#> #   ²​speechers, ³​meeting_unit
#>