Retrieving the records of parliamentary questions via Taiwan Legislative Yuan API 委員質詢事項

Retrieving the records of parliamentary questions via Taiwan Legislative Yuan API 委員質詢事項

Usage

get_parlquestions(term = 8, session_period = NULL, verbose = TRUE)

Arguments

term

Requesting questions from the term. The parameter should be set in a numeric format. The default value is 8. The data is only available from 8th term 參數必須為數值，資料從立法院第8屆開始計算。

session_period

session in the term. The session is between 1 and 8. session_period 參數必須為數值。

verbose

The default value is TRUE, displaying the description of data retrieved in number, url and computing time.

Value

A list object contains a tibble carrying

See also

Examples

## query parliamentary questions by term.
## 輸入「立委會期」下載立委質詢資料

get_parlquestions(term = 8)
#>  Retrieved URL: 
#>  https://data.ly.gov.tw/odw/ID6Action.action?term=08&sessionPeriod=&sessionTimes=&item=&fileType=json 
#>  Retrieved Term:  8 
#>  Retrieved Num:  4717 
#> $title
#> [1] "the records of parliarmentary questions"
#> 
#> $query_time
#> [1] "2023-01-27 20:10:38 GMT"
#> 
#> $retrieved_number
#> [1] 4717
#> 
#> $retrieved_term
#> [1] 8
#> 
#> $url
#> [1] "https://data.ly.gov.tw/odw/ID6Action.action?term=08&sessionPeriod=&sessionTimes=&item=&fileType=json"
#> 
#> $variable_names
#> [1] "term"          "sessionPeriod" "sessionTimes"  "item"         
#> [5] "selectTerm"   
#> 
#> $manual_info
#> [1] "https://data.ly.gov.tw/getds.action?id=6"
#> 
#> $data
#> # A tibble: 4,717 × 5
#>    term  sessionPeriod sessionTimes item                          select…¹
#>    <chr> <chr>         <chr>        <chr>                         <chr>   
#>  1 08    01            01           乙、本院委員質詢部分          all     
#>  2 08    01            01           一、本院楊委員瓊瓔，針對現行… all     
#>  3 08    01            01           二、本院楊委員瓊瓔，針對王金… all     
#>  4 08    01            01           三、本院楊委員瓊瓔，針對消費… all     
#>  5 08    01            01           四、本院楊委員瓊瓔，針對金融… all     
#>  6 08    01            01           五、本院楊委員瓊瓔，針對消基… all     
#>  7 08    01            01           六、本院楊委員瓊瓔，針對公用… all     
#>  8 08    01            01           七、本院楊委員瓊瓔，針對五都… all     
#>  9 08    01            01           八、本院楊委員瓊瓔，有鑑若干… all     
#> 10 08    01            01           九、本院楊委員瓊瓔，有鑑於往… all     
#> # … with 4,707 more rows, and abbreviated variable name ¹​selectTerm
#> 

## query parliamentary questions by term.
## 輸入「立委屆期」與「會期」下載立委質詢資料

get_parlquestions(term = 8, session_period = 2)
#>  Retrieved URL: 
#>  https://data.ly.gov.tw/odw/ID6Action.action?term=08&sessionPeriod=02&sessionTimes=&item=&fileType=json 
#>  Retrieved Term:  8 
#>  Retrieved Num:  923 
#> $title
#> [1] "the records of parliarmentary questions"
#> 
#> $query_time
#> [1] "2023-01-27 20:10:44 GMT"
#> 
#> $retrieved_number
#> [1] 923
#> 
#> $retrieved_term
#> [1] 8
#> 
#> $url
#> [1] "https://data.ly.gov.tw/odw/ID6Action.action?term=08&sessionPeriod=02&sessionTimes=&item=&fileType=json"
#> 
#> $variable_names
#> [1] "term"          "sessionPeriod" "sessionTimes"  "item"         
#> [5] "selectTerm"   
#> 
#> $manual_info
#> [1] "https://data.ly.gov.tw/getds.action?id=6"
#> 
#> $data
#> # A tibble: 923 × 5
#>    term  sessionPeriod sessionTimes item                           selec…¹
#>    <chr> <chr>         <chr>        <chr>                          <chr>  
#>  1 08    02            01           乙、本院委員質詢部分           all    
#>  2 08    02            01           一、本院林委員滄敏，鑒於國立…  all    
#>  3 08    02            01           二、本院何委員欣純，為因應全…  all    
#>  4 08    02            01           三、本院何委員欣純，基於節能…  all    
#>  5 08    02            01           四、本院何委員欣純，為維護嬰…  all    
#>  6 08    02            01           五、本院何委員欣純，針對台中…  all    
#>  7 08    02            01           六、本院何委員欣純，針對中興…  all    
#>  8 08    02            01           七、本院何委員欣純，針對台中…  all    
#>  9 08    02            01           八、本院楊委員曜，針對101年度… all    
#> 10 08    02            01           九、本院魏委員明谷，針對經濟…  all    
#> # … with 913 more rows, and abbreviated variable name ¹​selectTerm
#> 

#' ## query parliamentary questions by term.
## 輸入「空白」下載立委全部質詢資料

get_parlquestions(term = 8, session_period = 2)
#>  Retrieved URL: 
#>  https://data.ly.gov.tw/odw/ID6Action.action?term=08&sessionPeriod=02&sessionTimes=&item=&fileType=json 
#>  Retrieved Term:  8 
#>  Retrieved Num:  923 
#> $title
#> [1] "the records of parliarmentary questions"
#> 
#> $query_time
#> [1] "2023-01-27 20:10:46 GMT"
#> 
#> $retrieved_number
#> [1] 923
#> 
#> $retrieved_term
#> [1] 8
#> 
#> $url
#> [1] "https://data.ly.gov.tw/odw/ID6Action.action?term=08&sessionPeriod=02&sessionTimes=&item=&fileType=json"
#> 
#> $variable_names
#> [1] "term"          "sessionPeriod" "sessionTimes"  "item"         
#> [5] "selectTerm"   
#> 
#> $manual_info
#> [1] "https://data.ly.gov.tw/getds.action?id=6"
#> 
#> $data
#> # A tibble: 923 × 5
#>    term  sessionPeriod sessionTimes item                           selec…¹
#>    <chr> <chr>         <chr>        <chr>                          <chr>  
#>  1 08    02            01           乙、本院委員質詢部分           all    
#>  2 08    02            01           一、本院林委員滄敏，鑒於國立…  all    
#>  3 08    02            01           二、本院何委員欣純，為因應全…  all    
#>  4 08    02            01           三、本院何委員欣純，基於節能…  all    
#>  5 08    02            01           四、本院何委員欣純，為維護嬰…  all    
#>  6 08    02            01           五、本院何委員欣純，針對台中…  all    
#>  7 08    02            01           六、本院何委員欣純，針對中興…  all    
#>  8 08    02            01           七、本院何委員欣純，針對台中…  all    
#>  9 08    02            01           八、本院楊委員曜，針對101年度… all    
#> 10 08    02            01           九、本院魏委員明谷，針對經濟…  all    
#> # … with 913 more rows, and abbreviated variable name ¹​selectTerm
#>