Retrieving the records of national public debates 國是論壇

Retrieving the records of national public debates 國是論壇

get_public_debates(term = NULL, session_period = NULL, verbose = TRUE)

Arguments

term

Requesting answered questions by term. The parameter should be set in a numeric format. The default value is 8. The data is only available from 8th term 參數必須為數值，資料從立法院第8屆開始計算。

session_period

session in the term. The session is between 1 and 8. session_period 參數必須為數值。

verbose

The default value is TRUE, displaying the description of data retrieved in number, url and computing time.

Value

A list object contains a tibble carrying the variables of term, sessionPeriod, sessionTimes, meetingTimes, eyNumber, lyNumber, subject, content, docUrl selectTerm.

See also

Examples

## query the Executives' answered response by term and the session period.
## 輸入「立委屆期」與「會期」下載「質詢事項 (行政院答復部分)」
get_public_debates(term = 10, session_period = 2)
#>  Retrieved URL: 
#>  https://data.ly.gov.tw/odw/ID7Action.action?term=10&sessionPeriod=02&sessionTimes=&meetingTimes=&legislatorName=&speakType=&fileType=json 
#>  Retrieved Term:  10 
#>  Retrieved Num:  39 
#> $title
#> [1] "the records of the questions answered by the executives"
#> 
#> $query_time
#> [1] "2023-01-27 02:14:35 GMT"
#> 
#> $retrieved_number
#> [1] 39
#> 
#> $retrieved_term
#> [1] 10
#> 
#> $url
#> [1] "https://data.ly.gov.tw/odw/ID7Action.action?term=10&sessionPeriod=02&sessionTimes=&meetingTimes=&legislatorName=&speakType=&fileType=json"
#> 
#> $variable_names
#>  [1] "term"           "sessionPeriod"  "sessionTimes"   "meetingTimes"  
#>  [5] "dateTimeDesc"   "meetingRoom"    "chairman"       "legislatorName"
#>  [9] "speakType"      "content"        "selectTerm"    
#> 
#> $manual_info
#> [1] "https://data.ly.gov.tw/getds.action?id=7"
#> 
#> $data
#> # A tibble: 39 × 11
#>    term  sessi…¹ sessi…² meeti…³ dateT…⁴ meeti…⁵ chair…⁶ legis…⁷ speak…⁸ content
#>    <chr> <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#>  1 10    02      02      null    中華民… 本院議… 游錫堃  葉毓蘭  speak   （9時… 
#>  2 10    02      02      null    中華民… 本院議… 游錫堃  洪孟楷  speak   （9時4…
#>  3 10    02      02      null    中華民… 本院議… 游錫堃  孔文吉  speak   （9時7…
#>  4 10    02      02      null    中華民… 本院議… 游錫堃  陳以信  speak   （9時1…
#>  5 10    02      02      null    中華民… 本院議… 游錫堃  李德維  speak   （9時1…
#>  6 10    02      02      null    中華民… 本院議… 游錫堃  李貴敏  speak   （9時1…
#>  7 10    02      02      null    中華民… 本院議… 游錫堃  林思銘  speak   （9時2…
#>  8 10    02      02      null    中華民… 本院議… 游錫堃  楊瓊瓔  speak   （9時2…
#>  9 10    02      02      null    中華民… 本院議… 游錫堃  溫玉霞  speak   （9時2…
#> 10 10    02      03      null    中華民… 本院議… 蔡其昌  李德維  speak   （9時1…
#> # … with 29 more rows, 1 more variable: selectTerm <chr>, and abbreviated
#> #   variable names ¹​sessionPeriod, ²​sessionTimes, ³​meetingTimes, ⁴​dateTimeDesc,
#> #   ⁵​meetingRoom, ⁶​chairman, ⁷​legislatorName, ⁸​speakType
#> 

get_public_debates(term = 10, session_period = 1)
#>  Retrieved URL: 
#>  https://data.ly.gov.tw/odw/ID7Action.action?term=10&sessionPeriod=01&sessionTimes=&meetingTimes=&legislatorName=&speakType=&fileType=json 
#>  Retrieved Term:  10 
#>  Retrieved Num:  107 
#> $title
#> [1] "the records of the questions answered by the executives"
#> 
#> $query_time
#> [1] "2023-01-27 02:14:36 GMT"
#> 
#> $retrieved_number
#> [1] 107
#> 
#> $retrieved_term
#> [1] 10
#> 
#> $url
#> [1] "https://data.ly.gov.tw/odw/ID7Action.action?term=10&sessionPeriod=01&sessionTimes=&meetingTimes=&legislatorName=&speakType=&fileType=json"
#> 
#> $variable_names
#>  [1] "term"           "sessionPeriod"  "sessionTimes"   "meetingTimes"  
#>  [5] "dateTimeDesc"   "meetingRoom"    "chairman"       "legislatorName"
#>  [9] "speakType"      "content"        "selectTerm"    
#> 
#> $manual_info
#> [1] "https://data.ly.gov.tw/getds.action?id=7"
#> 
#> $data
#> # A tibble: 107 × 11
#>    term  sessi…¹ sessi…² meeti…³ dateT…⁴ meeti…⁵ chair…⁶ legis…⁷ speak…⁸ content
#>    <chr> <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#>  1 10    01      01      null    中華民… 本院議… 蔡其昌  江啟臣  speak   （9時1…
#>  2 10    01      01      null    中華民… 本院議… 蔡其昌  李德維  speak   （9時4…
#>  3 10    01      01      null    中華民… 本院議… 蔡其昌  楊瓊瓔  speak   （9時8…
#>  4 10    01      01      null    中華民… 本院議… 蔡其昌  邱臣遠  speak   （9時1…
#>  5 10    01      01      null    中華民… 本院議… 蔡其昌  李貴敏  speak   （9時1…
#>  6 10    01      01      null    中華民… 本院議… 蔡其昌  溫玉霞  speak   （9時1…
#>  7 10    01      01      null    中華民… 本院議… 蔡其昌  葉毓蘭  speak   （9時2…
#>  8 10    01      01      null    中華民… 本院議… 蔡其昌  洪孟楷  speak   （9時2…
#>  9 10    01      01      null    中華民… 本院議… 蔡其昌  吳斯懷  speak   （9時2…
#> 10 10    01      01      null    中華民… 本院議… 蔡其昌  吳怡玎  speak   （9時3…
#> # … with 97 more rows, 1 more variable: selectTerm <chr>, and abbreviated
#> #   variable names ¹​sessionPeriod, ²​sessionTimes, ³​meetingTimes, ⁴​dateTimeDesc,
#> #   ⁵​meetingRoom, ⁶​chairman, ⁷​legislatorName, ⁸​speakType
#>