The Records of National Public Debates 國是論壇

The Records of National Public Debates 國是論壇

Usage

get_public_debates(term = NULL, session_period = NULL, verbose = TRUE)

Arguments

term

numeric or NULL The default is set to 10. 參數必須為數值，資料從自 第8屆第1會期起，但實測資料從第10屆，故預設為10。

session_period

integer or NULL. Available options for the session periods is: 1, 2, 3, 4, 5, 6, 7, and 8. The default is set to NULL 參數必須為數值。

verbose

logical, indicates whether get_public_debates should print out detailed output when retrieving the data. The default is TRUE

Value

list, which contains:

title

the meeting records of cross-caucus session

query_time

the query time

retrieved_number

the number of observation

meeting_unit

the meeting unit

start_date_ad

the start date in POSIXct

end_date_ad

the end date in POSIXct

start_date

the start date in ROC Taiwan calendar

url

the retrieved json url

variable_names

the variables of the tibble dataframe

manual_info

the official manual, https://data.ly.gov.tw/getds.action?id=7; or use legisTaiwan::get_variable_info("get_public_debates")

data

a tibble dataframe, whose variables include:

sessionPeriod

會期

sessionTimes

會次

meetingTimes

臨時會會次

dateTimeDesc

日期時間說明

meetingRoom

會議地點

chairman

主持人

legislatorName

委員姓名

speakType

發言類型(paper:書面發言,speak:發言)

content

內容

selectTerm

屆別期別篩選條件

Details

get_public_debates produces a list, which contains title, query_time, retrieved_number, meeting_unit, start_date_ad, end_date_ad, start_date, end_date, url, variable_names, manual_info and data.

Note

To retrieve the user manual and more information about variable of the data frame, please use legisTaiwan::get_variable_info("get_public_debates") or visit the API manual at https://data.ly.gov.tw/getds.action?id=7. 議事類: 提供公報之國是論壇資訊，並包含書面意見。自第8屆第1會期起，但實測資料從第10屆。

See also

Examples

## query the Executives' answered response by term and the session period.
## 輸入「立委屆期」與「會期」下載國是論壇資訊。
get_public_debates(term = 10, session_period = 2)
#>  Retrieved URL: 
#>  https://data.ly.gov.tw/odw/ID7Action.action?term=10&sessionPeriod=02&sessionTimes=&meetingTimes=&legislatorName=&speakType=&fileType=json 
#>  Retrieved Term:  10 
#>  Retrieved Num:  39 
#> $title
#> [1] "the records of the questions answered by the executives"
#> 
#> $query_time
#> [1] "2023-02-11 09:40:53 CET"
#> 
#> $retrieved_number
#> [1] 39
#> 
#> $retrieved_term
#> [1] "10"
#> 
#> $url
#> [1] "https://data.ly.gov.tw/odw/ID7Action.action?term=10&sessionPeriod=02&sessionTimes=&meetingTimes=&legislatorName=&speakType=&fileType=json"
#> 
#> $variable_names
#>  [1] "term"           "sessionPeriod"  "sessionTimes"   "meetingTimes"  
#>  [5] "dateTimeDesc"   "meetingRoom"    "chairman"       "legislatorName"
#>  [9] "speakType"      "content"        "selectTerm"    
#> 
#> $manual_info
#> [1] "https://data.ly.gov.tw/getds.action?id=7"
#> 
#> $data
#> # A tibble: 39 × 11
#>    term  session…¹ sessi…² meeti…³ dateT…⁴ meeti…⁵ chair…⁶ legis…⁷ speak…⁸
#>    <chr> <chr>     <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#>  1 10    02        02      null    中華民… 本院議… 游錫堃  葉毓蘭  speak  
#>  2 10    02        02      null    中華民… 本院議… 游錫堃  洪孟楷  speak  
#>  3 10    02        02      null    中華民… 本院議… 游錫堃  孔文吉  speak  
#>  4 10    02        02      null    中華民… 本院議… 游錫堃  陳以信  speak  
#>  5 10    02        02      null    中華民… 本院議… 游錫堃  李德維  speak  
#>  6 10    02        02      null    中華民… 本院議… 游錫堃  李貴敏  speak  
#>  7 10    02        02      null    中華民… 本院議… 游錫堃  林思銘  speak  
#>  8 10    02        02      null    中華民… 本院議… 游錫堃  楊瓊瓔  speak  
#>  9 10    02        02      null    中華民… 本院議… 游錫堃  溫玉霞  speak  
#> 10 10    02        03      null    中華民… 本院議… 蔡其昌  李德維  speak  
#> # … with 29 more rows, 2 more variables: content <chr>, selectTerm <chr>,
#> #   and abbreviated variable names ¹​sessionPeriod, ²​sessionTimes,
#> #   ³​meetingTimes, ⁴​dateTimeDesc, ⁵​meetingRoom, ⁶​chairman,
#> #   ⁷​legislatorName, ⁸​speakType
#>