Retrieving full video information of meetings and committees 會議類:提供立法院院會及委員會之委員發言片段相關影片資訊。(自第9屆第1會期起)

Retrieving full video information of meetings and committees 會議類:提供立法院院會及委員會之委員發言片段相關影片資訊。(自第9屆第1會期起)

Usage

get_speech_video(start_date = NULL, end_date = NULL, verbose = TRUE)

Arguments

start_date

character Must be formatted in ROC Taiwan calendar with three forward slashes between year, month and day, e.g. "106/10/20".

end_date

character Must be formatted in ROC Taiwan calendar with three forward slashes between year, month and day, e.g. "109/01/10".

verbose

logical, indicates whether get_meetings should print out detailed output when retrieving the data.

Value

list, which contains:

title

the meeting records of cross-caucus session

query_time

the query time

retrieved_number

the number of observation

meeting_unit

the meeting unit

start_date_ad

the start date in POSIXct

end_date_ad

the end date in POSIXct

start_date

the start date in ROC Taiwan calendar

url

the retrieved json url

variable_names

the variables of the tibble dataframe

manual_info

the offical manual

data

a tibble dataframe, whose variables include: term: 屆期, sessionPeriod: 會期, meetingDate: 會議日期(西元年), meetingTime: 會議時間, meetingTypeName: 主辦單位, meetingName: 會議名稱, meetingContent:會議事由, legislatorName:委員姓名, areaName: 選區名稱, speechStartTime: 委員發言時間起, speechEndTime: 委員發言時間迄, speechRecordUrl: 發言紀錄網址, videoLength: 影片長度, videoUrl: 影片網址, and selectTerm:屆別期別篩選條件

Details

get_speech_video produces a list, which contains title, query_time, retrieved_number, meeting_unit, start_date_ad, end_date_ad, start_date, end_date, url, variable_names, manual_info and data.

See also

Examples

## query full video information of meetings and committees using a period of
## the dates in Taiwan ROC calender format with forward slash (/).
## 輸入「中華民國民年」下載「委員發言片段相關影片資訊」，輸入時間請依照該
## 格式 "105/10/20"，需有「正斜線」做隔開。
get_speech_video(start_date = "105/10/20", end_date = "109/03/10")
#>  Retrieved URL: 
#>  https://data.ly.gov.tw/odw/ID148Action.action?term=&sessionPeriod=&meetingDateS=105/10/20&meetingDateE=109/03/10&meetingTime=&legislatorName=&fileType=json 
#>  Retrieved date between: 2016-10-20 and 2020-03-10 
#>  Retrieved number: 547 
#> $title
#> [1] "the meeting records of cross-caucus session"
#> 
#> $query_time
#> [1] "2023-02-03 09:37:28 UTC"
#> 
#> $retrieved_number
#> [1] 547
#> 
#> $start_date_ad
#> [1] "2016-10-20"
#> 
#> $end_date_ad
#> [1] "2020-03-10"
#> 
#> $start_date
#> [1] "105/10/20"
#> 
#> $end_date
#> [1] "109/03/10"
#> 
#> $url
#> [1] "https://data.ly.gov.tw/odw/ID148Action.action?term=&sessionPeriod=&meetingDateS=105/10/20&meetingDateE=109/03/10&meetingTime=&legislatorName=&fileType=json"
#> 
#> $variable_names
#>  [1] "term"            "sessionPeriod"   "meetingDate"    
#>  [4] "meetingTime"     "meetingTypeName" "meetingName"    
#>  [7] "meetingContent"  "legislatorName"  "areaName"       
#> [10] "speechStartTime" "speechEndTime"   "speechRecordUrl"
#> [13] "videoLength"     "videoUrl"        "selectTerm"     
#> 
#> $manual_info
#> [1] "https://data.ly.gov.tw/getds.action?id=8"
#> 
#> $data
#> # A tibble: 547 × 15
#>    term  session…¹ meeti…² meeti…³ meeti…⁴ meeti…⁵ meeti…⁶ legis…⁷ areaN…⁸
#>    <chr> <chr>     <chr>   <chr>   <chr>   <chr>   <chr>   <chr>   <chr>  
#>  1 10    1         2020-0… 09:00   院會    第10屆… "21日…  李德維  全國不…
#>  2 10    1         2020-0… 09:00   院會    第10屆… "21日…  賴香伶  全國不…
#>  3 10    1         2020-0… 09:00   院會    第10屆… "21日…  賴瑞隆  高雄市…
#>  4 10    1         2020-0… 09:00   院會    第10屆… "21日…  鄭運鵬  桃園市…
#>  5 10    1         2020-0… 09:00   院會    第10屆… "21日…  吳斯懷  全國不…
#>  6 10    1         2020-0… 09:00   院會    第10屆… "21日…  邱顯智  全國不…
#>  7 10    1         2020-0… 09:00   院會    第10屆… "21日…  高虹安  全國不…
#>  8 10    1         2020-0… 09:00   院會    第10屆… "21日…  楊瓊瓔  台中市…
#>  9 10    1         2020-0… 09:00   院會    第10屆… "21日…  李貴敏  全國不…
#> 10 10    1         2020-0… 09:00   院會    第10屆… "21日…  陳以信  全國不…
#> # … with 537 more rows, 6 more variables: speechStartTime <chr>,
#> #   speechEndTime <chr>, speechRecordUrl <chr>, videoLength <chr>,
#> #   videoUrl <chr>, selectTerm <chr>, and abbreviated variable names
#> #   ¹​sessionPeriod, ²​meetingDate, ³​meetingTime, ⁴​meetingTypeName,
#> #   ⁵​meetingName, ⁶​meetingContent, ⁷​legislatorName, ⁸​areaName
#>